description
Today security is a main issue for protecting the resources. Security is important because risk of intrusion and theft has become increasing. Security is also necessary for protecting banks from fire and other abnormal activities. Many people are using various types of security systems. We have found that most of the security systems are developed only for alarm using microcontrollers; in our research we use Microcontrollers with different sensors (PIR, Smoke or Fire, IR and Gas) as observatory to detect or identify intruder or abnormal activities inside the bank and ATM. The main aim of this research is to design a system for alerting theft and to auto arrest the thief in bank or ATM itself from centralized monitoring unit. The purpose of the system is to design a smart and centralized monitoring and control system using IOT technologies. Bank Security System, IOT based security system, monitoring system, intruder detection system, web monitoring system.